change a code according to comments in the PR
