+++ src/pages/sharedNotes/styled.jsx
@@ -1,19 +1,4 @@
-import { Container } from '@mui/material';
 import { styled } from '@mui/styles';
-
-export const SharedBox = styled(Container)({
-    '&.MuiContainer-root': {
-        maxWidth: '100%',
-        width: '700px',
-        backgroundColor: 'white',
-        padding: '20px',
-        borderRadius: '5px',
-        marginTop: '50px',
-        display: 'flex',
-        flexDirection: 'column',
-        alignItems: 'center',
-    },
-});
 
 export const PageTitle = styled('h2')({
     fontSize: '36px',
@@ -21,6 +6,7 @@ export const PageTitle = styled('h2')({
     color: 'ghostwhite',
     textAlign: 'center',
     marginTop: '20px',
+    textTransform: 'uppercase',
     '&.dark': {
         color: 'black',
         marginTop: '0',
